parte 1

+++ src/main/java/edu/eci/cvds/sampleprj/dao/mybatis/mappers/ClienteMapper.java
@@ -22,7 +22,7 @@ public interface ClienteMapper {
      * @param fechainicio
      * @param fechafin 
      */
-    public void agregarItemRentadoACliente(@Param("id") int id, 
+    public void agregarItemRentadoACliente(@Param("id") long id, 
             @Param("idit") int idit, 
             @Param("fechaInicio") Date fechainicio,
             @Param("fechaFin") Date fechafin);
